[Refactor] HashTag -> HairStyle
