+++ src/main/java/com/inq/wishhair/wesharewishhair/hairstyle/controller/HairStyleController.java
@@ -3,7 +3,7 @@ package com.inq.wishhair.wesharewishhair.hairstyle.controller;
 import com.inq.wishhair.wesharewishhair.auth.config.resolver.ExtractPayload;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.hairstyle.service.HairStyleService;
-import com.inq.wishhair.wesharewishhair.hashtag.enums.Tag;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.enums.Tag;
 import com.inq.wishhair.wesharewishhair.hairstyle.controller.dto.response.HairStyleResponse;
 import com.inq.wishhair.wesharewishhair.hairstyle.controller.dto.response.PagedHairStyleResponse;
 import lombok.RequiredArgsConstructor;
