+++ src/main/java/com/inq/wishhair/wesharewishhair/hairstyle/domain/HairStyle.java
@@ -1,7 +1,7 @@
 package com.inq.wishhair.wesharewishhair.hairstyle.domain;
 
 
-import com.inq.wishhair.wesharewishhair.hashtag.HashTag;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.HashTag;
 import com.inq.wishhair.wesharewishhair.photo.entity.Photo;
 import com.inq.wishhair.wesharewishhair.user.enums.Sex;
 import jakarta.persistence.*;
